+++ Include/Library/OcBootManagementLib.h
@@ -101,6 +101,12 @@ typedef UINT32 OC_BOOT_ENTRY_TYPE;
 #define OC_BOOT_SYSTEM              (OC_BOOT_RESET_NVRAM)
 
 /**
+  Default boot option numbers.
+**/
+#define OC_BOOT_OPTION                0x9696
+#define OC_BOOT_OPTION_VARIABLE_NAME  L"Boot9696"
+
+/**
   Picker mode.
 **/
 typedef enum OC_PICKER_MODE_ {
@@ -1097,6 +1103,22 @@ VOID
 OcToggleVoiceOver (
   IN  OC_PICKER_CONTEXT  *Context,
   IN  UINT32             File  OPTIONAL
+  );
+
+/**
+  Register top-most priority boot option.
+
+  @param[in]  OptionName    Option name to create.
+  @param[in]  DeviceHandle  Device handle of the file system.
+  @param[in]  FilePath      Bootloader path.
+
+  @retval EFI_SUCCESS on success.
+**/
+EFI_STATUS
+OcRegisterBootOption (
+  IN CONST CHAR16    *OptionName,
+  IN EFI_HANDLE      DeviceHandle,
+  IN CONST CHAR16    *FilePath
   );
 
 #endif // OC_BOOT_MANAGEMENT_LIB_H