+++ Changelog.md
@@ -17,6 +17,7 @@ OpenCore Changelog
 - Added `AppleRtcRam` protocol implementation
 - Renamed `Protocols` to `ProtocolOverrides` for clarity
 - Added ResetSystem tool to allow shutdown/reset actions in the menu
+- Added experimental `BootProtect` `Security` option
 
 #### v0.5.7
 - Added TimeMachine detection to picker